requirements.txt
# Core LangChain packages
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.3.0

# Core dependencies for PDF Chat app
streamlit>=1.22.0
openai>=1.0.0
chromadb>=0.3.21
pypdf>=3.8.1
pdf2image>=1.16.3
Pillow>=9.5.0
python-dotenv>=1.0.0
